+++ pkg/azuremachinepool/mutate_azuremachinepool_create.go
@@ -10,6 +10,7 @@ import (
 	expcapzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
 
 	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
+	"github.com/giantswarm/azure-admission-controller/pkg/generic"
 	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
 )
 
@@ -52,7 +53,15 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		return []mutator.PatchOperation{}, microerror.Maskf(parsingFailedError, "unable to parse azureMachinePool CR: %v", err)
 	}
 
-	patch, err := m.ensureStorageAccountType(ctx, azureMPCR)
+	patch, err := generic.EnsureOrganizationLabelNormalized(ctx, azureMPCR)
+	if err != nil {
+		return []mutator.PatchOperation{}, microerror.Mask(err)
+	}
+	if patch != nil {
+		result = append(result, *patch)
+	}
+
+	patch, err = m.ensureStorageAccountType(ctx, azureMPCR)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}