+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- AzureCluster, AzureMachine, AzureMachinePool, Cluster and MachinePools CR's
+  are ensured to have normalized form of giantswarm.io/organization label
+  value via mutating webhook on CREATE.
+
 ### Changed
 
 - Set `400` status code if a request is invalid.